refactor: extract ObservableObjectChangeForwarder.swift from AppPresentationState.swift
Splits the `ObservableObjectChangeForwarder` helper (Combine plumbing
that fans a child ObservableObject's changes into a parent's
objectWillChange) out of AppPresentationState.swift into its own file.

Byte-identical move of lines 5-20. New file needs only Combine
(no AppKit); AppPresentationState.swift keeps AppKit + Combine +
Foundation for AppLifecycleNotificationBinder + AppPresentationState.

AppPresentationState.swift: 76 → 59 lines. Tests: 667.
